+++ handler.ts
@@ -4,4 +4,4 @@ export { default as auth } from "./src/auth";
 export { default as featureFlag } from "./src/featureFlag";
 export { default as manualMatch } from "./src/manualMatch";
 export { default as handleMatch } from "./src/handleMatch";
-export { default as createMatch } from "./src/createMatch";
+export { default as createNewMatch } from "./src/createNewMatch";
